requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
apscheduler==3.10.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.3.0
httpx==0.25.2
aiofiles==23.2.1
